+++ example/github/remote-data-blocks/inc/patterns/file-render.html
@@ -1,0 +1,7 @@
+<!-- wp:group {"layout":{"type":"constrained"}} -->
+<div class="wp-block-group github-file-content">
+	<!-- wp:paragraph {"metadata":{"bindings":{"content":{"source":"remote-data/binding","args":{"field":"file_content"}}},"name":"File Content"}} -->
+	<p></p>
+	<!-- /wp:paragraph -->
+</div>
+<!-- /wp:group -->
