--- example/github/remote-data-blocks/inc/php-markdown/.editorconfig
@@ -1,11 +1,0 @@
-# EditorConfig is awesome: https://EditorConfig.org
-root = true
-
-[*]
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = false
-
-[*.php]
-indent_style = tab
-insert_final_newline = true
