--- example/github/remote-data-blocks/inc/php-markdown/.gitignore
@@ -1,3 +1,0 @@
-*~
-/composer.lock
-/vendor/
