--- example/github/remote-data-blocks/inc/php-markdown/Michelf/Markdown.inc.php
@@ -1,10 +1,0 @@
-<?php
-
-// Use this file if you cannot use class autoloading. It will include all the
-// files needed for the Markdown parser.
-//
-// Take a look at the PSR-0-compatible class autoloading implementation
-// in the Readme.php file if you want a simple autoloader setup.
-
-require_once dirname(__FILE__) . '/MarkdownInterface.php';
-require_once dirname(__FILE__) . '/Markdown.php';
