+++ example/github/remote-data-blocks/inc/queries/class-github-datasource.php
@@ -5,14 +5,14 @@ namespace RemoteDataBlocks\Example\GitHub;
 use RemoteDataBlocks\Config\HttpDatasource;
 
 class GitHubDatasource extends HttpDatasource {
-	private string $repo_owner;
-	private string $repo_name;
-	private string $branch;
+	public string $repo_owner;
+	public string $repo_name;
+	public string $ref;
 
-	public function __construct( string $repo_owner, string $repo_name, string $branch = 'trunk' ) {
+	public function __construct( string $repo_owner, string $repo_name, string $ref = 'main' ) {
 		$this->repo_owner = $repo_owner;
 		$this->repo_name  = $repo_name;
-		$this->branch     = $branch;
+		$this->ref        = $ref;
 	}
 
 	public function get_endpoint(): string {
@@ -20,7 +20,7 @@ class GitHubDatasource extends HttpDatasource {
 			'https://api.github.com/repos/%s/%s/git/trees/%s?recursive=1',
 			$this->repo_owner,
 			$this->repo_name,
-			$this->branch
+			$this->ref
 		);
 	}
