+++ example/github/remote-data-blocks/inc/queries/class-github-list-files-query.php
