+++ example/github/remote-data-blocks/register.php
@@ -7,8 +7,8 @@ use RemoteDataBlocks\Logging\LoggerManager;
 use function add_action;
 
 require_once __DIR__ . '/inc/queries/class-github-datasource.php';
-require_once __DIR__ . '/inc/queries/class-github-get-list-files-query.php';
-require_once __DIR__ . '/inc/queries/class-github-get-raw-file-query.php';
+require_once __DIR__ . '/inc/queries/class-github-get-file-as-html-query.php';
+require_once __DIR__ . '/inc/queries/class-github-list-files-query.php';
 
 function register_github_file_block() {
 	$repo_owner  = 'Automattic';
@@ -18,15 +18,16 @@ function register_github_file_block() {
 	$block_name  = sprintf( 'GitHub File (%s/%s)', $repo_owner, $repo_name );
 
 	$github_datasource             = new GitHubDatasource( $repo_owner, $repo_name, $branch );
+	$github_get_file_as_html_query = new GitHubGetFileAsHtmlQuery( $github_datasource );
 	$github_get_list_files_query   = new GitHubListFilesQuery( $github_datasource );
-	$github_get_raw_file_query     = new GitHubGetRawFileQuery( $github_datasource );
 
-    ConfigurationLoader::register_block( $block_name, $github_get_raw_file_query );
+    ConfigurationLoader::register_block( $block_name, $github_get_file_as_html_query );
 	ConfigurationLoader::register_list_query( $block_name, $github_get_list_files_query );
-	// ConfigurationLoader::register_page( $block_name, 'github-remote-data-blocks-embed-file' );
 
-	$block_pattern = file_get_contents( __DIR__ . '/inc/patterns/file-picker.html' );
-	ConfigurationLoader::register_block_pattern( $block_name, 'remote-data-blocks/github-file-picker', $block_pattern, [ 'title' => 'GitHub File Picker' ] );
+	$block_pattern1 = file_get_contents( __DIR__ . '/inc/patterns/file-picker.html' );
+	$block_pattern2 = file_get_contents( __DIR__ . '/inc/patterns/file-render.html' );
+	ConfigurationLoader::register_block_pattern( $block_name, 'remote-data-blocks/github-file-picker', $block_pattern1, [ 'title' => 'GitHub File Picker' ] );
+	ConfigurationLoader::register_block_pattern( $block_name, 'remote-data-blocks/github-file-render', $block_pattern2, [ 'title' => 'GitHub File Render' ] );
 
 	$logger = LoggerManager::instance();
 	$logger->info( sprintf( 'Registered %s block (branch: %s)', $block_name, $branch ) );
