+++ inc/config/query-runner/query-runner.php
@@ -11,7 +11,6 @@ namespace RemoteDataBlocks\Config;
 
 use Exception;
 use GuzzleHttp\RequestOptions;
-use JsonPath\JsonObject;
 use RemoteDataBlocks\HttpClient;
 use RemoteDataBlocks\Logging\LoggerManager;
 use WP_Error;
@@ -108,75 +107,12 @@ class QueryRunner implements QueryRunnerInterface {
 			$logger->warning( sprintf( 'Query error: %s', esc_html( $response_data['errors'][0]['message'] ) ) );
 		}
 
-		// This method always returns an array, even if it's a single item. This
-		// ensures a consistent response shape. The requestor is expected to inspect
-		// is_collection and unwrap if necessary.
-		$results = $this->map_fields( $response_data, $is_collection );
+		$results = $this->query_context->get_results( $response_data, $input_variables );
 
 		return [
 			'is_collection' => $is_collection,
 			'metadata'      => $this->query_context->get_metadata( $response, $results ),
 			'results'       => $results,
 		];
-	}
-
-	private function get_field_value( array|string $field_value, string $default_value = '', string $field_type = 'string' ): string {
-		$field_value_single = is_array( $field_value ) && count( $field_value ) > 1
-			? $field_value
-			: ( $field_value[0] ?? $default_value );
-
-		switch ( $field_type ) {
-			case 'price':
-				return sprintf( '$%s', number_format( $field_value_single, 2 ) );
-
-			case 'string':
-				return wp_strip_all_tags( $field_value_single );
-		}
-
-		return $field_value_single;
-	}
-
-	private function map_fields( $response_data, $is_collection = false ): array|null {
-		$root             = $response_data;
-		$output_variables = $this->query_context->output_variables;
-
-		if ( ! empty( $output_variables['root_path'] ) ) {
-			$json = new JsonObject( $root );
-			$root = $json->get( $output_variables['root_path'] );
-		} else {
-			$root = $is_collection ? $root : [ $root ];
-		}
-
-		if ( empty( $root ) || empty( $output_variables['mappings'] ) ) {
-			return $root;
-		}
-
-		// Loop over the returned items in the query result.
-		return array_map( function ( $item ) use ( $output_variables ) {
-			$json = new JsonObject( $item );
-
-			// Loop over the output variables and extract the values from the item.
-			$result = array_map( function ( $mapping ) use ( $json ) {
-				if ( array_key_exists( 'generate', $mapping ) && is_callable( $mapping['generate'] ) ) {
-					$field_value_single = $mapping['generate']( json_decode( $json->getJson(), true ) );
-				} else {
-					$field_path  = $mapping['path'] ?? null;
-					$field_value = $field_path ? $json->get( $field_path ) : '';
-
-					// JSONPath always returns values in an array, even if there's only one value.
-					// Because we're mostly interested in single values for field mapping, unwrap the array if it's only one item.
-					$field_value_single = self::get_field_value( $field_value, $mapping['defaultValue'] ?? '', $mapping['type'] );
-				}
-
-				return array_merge( $mapping, [
-					'value' => $field_value_single,
-				] );
-			}, $output_variables['mappings'] );
-
-			// Nest result property to reserve additional meta in the future.
-			return [
-				'result' => $result,
-			];
-		}, $root );
-	}
+	}	
 }
